Only start gb in rungb.jl script
